Create Color Package And Moved All Files There, Added GitHub Action For Deployment
