+++ .github/workflows/package.cd.yaml
@@ -1,0 +1,34 @@
+name: Package CD
+
+on:
+  release:
+    types:
+      - released
+
+jobs:
+  publish:
+    name: Publish Package
+    runs-on: ubuntu-latest
+    steps:
+      - name: Check Out
+        uses: actions/checkout@v4
+
+      - name: Set Up JDK 21 & Maven Central
+        uses: actions/setup-java@v4
+        with:
+          java-version: "21"
+          distribution: temurin
+          cache: maven
+          server-id: central
+          server-username: MAVEN_USERNAME
+          server-password: MAVEN_PASSWORD
+          gpg-private-key: GPG_KEY
+          gpg-passphrase: GPG_PASSPHRASE
+
+      - name: Publish Package To Maven Central
+        run: mvn --batch-mode clean deploy
+        env:
+          MAVEN_USERNAME: ${{ secrets.MAVEN_USERNAME }}
+          MAVEN_PASSWORD: ${{ secrets.MAVEN_PASSWORD }}
+          GPG_KEY: ${{secrets.GPG_KEY}}
+          GPG_PASSPHRASE: ${{secrets.GPG_PASSPHRASE}}
