+++ src/main/java/best/skn/utils/color/ColorName.java
@@ -1,114 +1,114 @@
-package best.skn.utils;
+package best.skn.utils.color;
 
 /**
  * ColorName Enum for internal use in Color Class
  *
  * @author SKN
- * @version 1.0.0
- * @since 26-02-2024
- * @dedicated.to Trina, Logno And Atoshi
+ * @version 1.1.0
+ * @since 07-03-2024
+ * @dedicated.to Trina, Logno and Atoshi
  */
 public enum ColorName {
   /**
    * Reset element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   RESET,
   /**
    * Black element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   BLACK,
   /**
    * Red element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   RED,
   /**
    * Green element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   GREEN,
   /**
    * Yellow element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   YELLOW,
   /**
    * Blue element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   BLUE,
   /**
    * Purple element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   PURPLE,
   /**
    * Cyan element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   CYAN,
   /**
    * White element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   WHITE,
   /**
    * Bold Black element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   BLACK_BOLD,
   /**
    * Bold Red element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   RED_BOLD,
   /**
    * Bold Green element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   GREEN_BOLD,
   /**
    * Bold Yellow element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   YELLOW_BOLD,
   /**
    * Bold Blue element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   BLUE_BOLD,
   /**
    * Bold Purple element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   PURPLE_BOLD,
   /**
    * Bold Cyan element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   CYAN_BOLD,
   /**
    * Bold White element
    *
-   * @since v1.0.0
+   * @since v1.1.0
    */
   WHITE_BOLD,
 }
