+++ src/main/java/best/skn/utils/color/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Utility Color Package for this Java Library
+ *
+ * @author SKN
+ * @version 1.1.0
+ * @since 07-03-2024
+ * @dedicated.to Trina, Logno and Atoshi
+ */
+package best.skn.utils.color;
