+++ src/main/java/best/skn/utils/package-info.java
@@ -4,6 +4,6 @@
  * @author SKN
  * @version 1.0.0
  * @since 26-02-2024
- * @dedicated.to Trina, Logno And Atoshi
+ * @dedicated.to Trina, Logno and Atoshi
  */
 package best.skn.utils;
